Clean up a little more
