+++ integration-tester/cmd/run.go
@@ -1,8 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-
 	"github.com/abetterinternet/prio-server/integration-tester/tester"
 	"github.com/spf13/cobra"
 )
@@ -11,9 +9,8 @@ var runCmd = &cobra.Command{
 	Use:   "run",
 	Short: "Run the integration-tester",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		fmt.Println(namespace, name, ownManifestUrl, phaManifestUrl, facilManifestUrl, serviceAccountName, facilitatorImage, pushGateway, awsAccountId)
 		t := tester.New(kubeConfigPath,
-			namespace, name,
+			namespace, ingestorLabel,
 			ownManifestUrl, phaManifestUrl, facilManifestUrl,
 			serviceAccountName, facilitatorImage,
 			pushGateway, awsAccountId,
